% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-type-specific_net.R
\name{calcModuleScore}
\alias{calcModuleScore}
\title{Calculate Module-Specific Scores for a Given Cell Type}
\usage{
calcModuleScore(count, celltype_module, celltye_name, meta)
}
\arguments{
\item{count}{A matrix of gene expression counts with genes as rows and samples as columns.}

\item{celltype_module}{The clustering result is obtained by selectCellModules function. 
for the specified cell type.}

\item{celltye_name}{A character string specifying the cell type to calculate scores for.}

\item{meta}{A metadata data frame with cell type information for each sample.}
}
\value{
A list with `M1_expr` and `M2_expr`, which represent the average expression sums for modules 1 and 2, respectively.
}
\description{
This function calculates module-specific scores (expression sums) for two gene modules
within a specific cell type, using normalized counts.
}
