% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-type-specific_net.R
\name{CelltypeClusterRes}
\alias{CelltypeClusterRes}
\title{Identify Cell Type-Specific Modules Based on Expression Scores}
\usage{
CelltypeClusterRes(M1_expr, M2_expr, Cluster_res, celltype_cor)
}
\arguments{
\item{M1_expr}{Numeric; the average expression score for module 1 within the specified cell type.}

\item{M2_expr}{Numeric; the average expression score for module 2 within the specified cell type.}

\item{Cluster_res}{Data frame; initial clustering results for genes associated with the specified cell type. 
This data frame includes gene identifiers and their assigned clusters.}

\item{celltype_cor}{Matrix; a correlation matrix of gene expression specific to the selected cell type. 
Each entry represents the correlation between gene pairs.}
}
\value{
A refined clustering result (`Cluster_res`) for the selected module, which represents the cell type-specific gene network.
}
\description{
This function identifies the most relevant module for a specified cell type based on module-specific expression scores. 
It selects either module 1 or module 2 based on their average expression scores, and further clusters 
the associated genes for the chosen module to refine the cell type-specific network.
}
