% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-type-specific_net.R
\name{CelltypeFeat}
\alias{CelltypeFeat}
\title{Cell Type-Specific Gene Co-expression Network}
\usage{
CelltypeFeat(DEG_path, feat, count, sort_meta)
}
\arguments{
\item{DEG_path}{Marker genes for different cell types.}

\item{feat}{A feat matrix from GeneCluster model.}

\item{count}{Raw count matrix.}
}
\description{
Cell Type-Specific Gene Co-expression Network
}
\details{
This module contains functions for constructing cell type-specific gene co-expression networks.
}
\author{
Byting
}
