% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_coexpression_net.R
\name{CelltypeModuleMap}
\alias{CelltypeModuleMap}
\title{Function to associate modules with cell types}
\usage{
CelltypeModuleMap(clusterRes, marker)
}
\arguments{
\item{clusterRes}{Global cluster res.}

\item{marker}{Celltype marker gene.}
}
\value{
Celltype-module assocation matrix and P-value.
}
\description{
This function associates global gene modules with cell types.
}
