% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-type-specific_net.R
\name{CelltypeNet}
\alias{CelltypeNet}
\title{Cell-type-specific network}
\usage{
CelltypeNet(celltype_feat_path, celltye_name, count, meta)
}
\arguments{
\item{celltype_feat_path}{A character string specifying the file path to the cell type-specific feature data in CSV format.
Each row represents a gene, and the columns represent features or conditions relevant to the cell type.}

\item{celltye_name}{A character string specifying the name of the cell type being analyzed.}

\item{count}{A matrix of gene expression counts with genes as rows and samples as columns.
This matrix provides the expression data for calculating module-specific scores.}

\item{meta}{A data frame containing metadata information for each sample, such as cell type labels.}
}
\value{
A data frame representing the refined clustering result for the selected module within the specified cell type.
Each row corresponds to a gene and includes clustering information specific to the cell type.
}
\description{
This function obtains cell type specific clustering results.
}
\examples{
# Example usage:
# CelltypeNet("data/celltype_features.csv", "B cell", count_matrix, meta_data)

}
