% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-type-specific_net.R
\name{compute_cosine_cormat}
\alias{compute_cosine_cormat}
\title{Compute Cosine Similarity Correlation Matrices}
\usage{
compute_cosine_cormat(input_dir, output_dir)
}
\arguments{
\item{input_dir}{Path to the input directory containing feature CSV files.}

\item{output_dir}{Path to the output directory where correlation matrices will be saved.}
}
\value{
A list of correlation matrices (one for each input file).
}
\description{
This function reads feature matrices (CSV files) from an input directory,
computes pairwise cosine similarity for each file, 
and saves the correlation matrices as CSV files in the output directory.
}
\examples{
\dontrun{
compute_cosine_cormat(
  input_dir = "hESC_GCN/time_feat/v2/",
  output_dir = "hESC_GCN/time_cormat/v2/"
)
}
}
