% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{findHVG_py}
\alias{findHVG_py}
\title{Find Highly Variable Genes using Python (Scanpy)}
\usage{
findHVG_py(expr, num_HVG, python_env = NULL)
}
\arguments{
\item{expr}{A data frame or matrix (genes as columns, cells as rows).}

\item{num_HVG}{Integer number of highly variable genes to return.}

\item{python_env}{Path to the Python environment containing scanpy and numpy.}
}
\value{
A data frame containing only the HVGs.
}
\description{
This function calls a Python implementation of HVG selection via \code{scanpy},
using \code{reticulate} to run Python code within R.
}
\examples{
\dontrun{
  res <- findHVG_py(expr, num_HVG = 1000, python_env = "/path/to/conda_env")
}
}
