% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_coexpression_net.R
\name{global_net}
\alias{global_net}
\title{Gene Co-expression Network}
\usage{
global_net(feat_matrix)
}
\arguments{
\item{cor}{A frature matrix from GeneCluster model.}
}
\value{
Data frames of cormatrix and global cluster.
}
\description{
Gene Co-expression Network Construction
}
\details{
This module contains functions for constructing gene co-expression networks.
}
\author{
Byting
}
