% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-type-specific_net.R
\name{heatmap_plot}
\alias{heatmap_plot}
\title{Heatmap Plotting}
\usage{
heatmap_plot(cor_matrix, marker)
}
\arguments{
\item{cor_matrix}{A correlation matrix.}

\item{marker}{A data frame containing marker genes.}
}
\value{
A heatmap object.
}
\description{
This function generates a heatmap for the correlation matrix.
}
