% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_coexpression_net.R
\name{hist_enrichment_plot}
\alias{hist_enrichment_plot}
\title{Histogram Enrichment Plot}
\usage{
hist_enrichment_plot(pvalue, cross_table)
}
\arguments{
\item{pvalue}{A matrix of p-values.}

\item{cross_table}{A matrix of cross-table data.}
}
\value{
A heatmap object.
}
\description{
This function creates a heatmap to visualize p-values and cross-table data.
}
