% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-type-specific_net.R
\name{marker_identify}
\alias{marker_identify}
\title{Cell-Type-Specific Gene Co-expression Network}
\usage{
marker_identify(expr, metadata_path, output_dir)
}
\arguments{
\item{expr}{A numeric matrix or data frame of gene expression (genes x cells).}

\item{metadata_path}{Path to a CSV file containing metadata.
The row names should match the column names of `expr`.
Must contain a column named `time_info` for grouping.}

\item{output_dir}{Path to save the marker genes CSV file.}
}
\value{
A data frame containing marker genes identified by Seurat.
}
\description{
This package provides functions to construct and analyze cell-type-specific networks.
Identify Marker Genes Using Seurat
}
\details{
This function identifies marker genes for each group in the metadata
using Seurat's FindAllMarkers function.
}
\examples{
\dontrun{
markers <- marker_identify(expr, "meta.csv", "markers.csv")
}

}
