% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{process_expression_data}
\alias{process_expression_data}
\title{Data Preprocess}
\usage{
process_expression_data(input_file, num_HVG)
}
\arguments{
\item{input_file}{A string representing the path to the input file.}

\item{num_HVG}{An integer specifying the number of highly variable genes to find.}
}
\value{
A data frame containing the normalized expression data.
}
\description{
Data Preprocessing
}
\details{
This script is used to read and process expression data.
}
\author{
Byting
}
