% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-type-specific_net.R
\name{Run_GO}
\alias{Run_GO}
\title{Run GO Enrichment Analysis}
\usage{
Run_GO(Cluster_res)
}
\arguments{
\item{Cluster_res}{A data frame containing cluster results.}

\item{universe}{A character vector of universe genes.}
}
\value{
A data frame of GO enrichment results.
}
\description{
This function performs GO enrichment analysis on clustered genes.
}
