% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-type-specific_net.R
\name{second_cluster}
\alias{second_cluster}
\title{Cluster Using Dynamic Tree Cut}
\usage{
second_cluster(cor, minClusterSize = 10)
}
\arguments{
\item{cor}{A correlation matrix.}
}
\value{
A data frame of clusters.
}
\description{
This function performs clustering on a correlation matrix using dynamic tree cut.
}
