% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-type-specific_net.R
\name{selectCellModules}
\alias{selectCellModules}
\title{Perform First Clustering}
\usage{
selectCellModules(celltype_feat)
}
\arguments{
\item{celltype_feat}{Selected cell type features.}

\item{cor_matrix}{A correlation matrix of genes.}
}
\value{
A data frame with gene clusters. Each gene is assigned a module number.
}
\description{
Perform First Clustering
}
