% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-type-specific_net.R
\name{CelltypeFeat}
\alias{CelltypeFeat}
\title{Get the Top 10% Features for Each Cell Type}
\usage{
CelltypeFeat(DEG, feat, expr, meta, CellType_col)
}
\arguments{
\item{feat}{A feat matrix from GeneCluster model.}

\item{expr}{expression matrix.}

\item{DEG_path}{Marker genes for different cell types.}
}
\description{
Get the Top 10% Features for Each Cell Type
}
